合并了用户管理和权限管理模块；统一了用户和角色表的crud逻辑，创建、删除、修改用户角色时 会对应更新角色表。

+++ src/main/java/com/hospital/repository/NurseRepository.java
@@ -4,7 +4,7 @@ import com.hospital.entity.Nurse;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface NurseRepository extends JpaRepository<Nurse, Integer> {
-    // 根据用户ID查找护士信息
-    Optional<Nurse> findByUserId(Integer userId);
+public interface NurseRepository extends JpaRepository<Nurse, Long> {
+    // 根据用户 ID 查找护士信息
+    Optional<Nurse> findByUserId(Long userId);
 }
